Leave middleware out of JoiQL core
