--- .env.example
@@ -1,6 +1,0 @@
-BLUEBIRD_LONG_STACK_TRACES=true
-BLUEBIRD_DEBUG=true
-ARTSY_URL=https://stagingapi.artsy.net
-POSITRON_URL=https://writer.artsy.net
-ARTSY_ID=
-ARTSY_SECRET=
