+++ test/integration.js
@@ -22,31 +22,27 @@ const Person = object({
   birthday: date(),
   age: number().integer()
 }).meta({
-  args: { id: string().required(), age: number().min(1).max(100) }
+  args: { id: string().required(), age: number().min(1).max(100) },
+  resolve: (source, args) => db[args.id]
 })
 
-const api = joiql({
+const schema = joiql({
   query: {
     person: Person
   }
 })
 
-api.use((ctx, next) => {
-  ctx.res.person = db[ctx.req.query.person.args.id]
-  return next()
-})
-
 describe('joiql', () => {
-  it('converts a Joi schema and middleware into GraphQL', () => {
+  it('converts a Joi schema into GraphQL', () => {
     const query = '{ person(id: "hillary") { name } }'
-    return graphql(api.schema, query).then((res) => {
+    return graphql(schema, query).then((res) => {
       res.data.person.name.should.equal('Hillary Clinton')
     })
   })
 
   it('validates args', () => {
     const query = '{ person(age: 0 id: "hillary") { name } }'
-    return graphql(api.schema, query).then((res) => {
+    return graphql(schema, query).then((res) => {
       res.errors[0].message.should.containEql('child "age" fails')
     })
   })
